requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asyncpg==0.29.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.115.0
googleapis-common-protos==1.59.1
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
idna==3.10
importlib_metadata==8.4.0
opentelemetry-api==1.27.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-asyncpg==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
protobuf==4.25.5
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyodbc==5.1.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
thrift==0.21.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
wrapt==1.16.0
zipp==3.20.2
